Mise à jour 2 Models, controllers et migrations
